#55 font系を font_control.py に分離した。
